+++ ngraph/ngraph.h
@@ -15,19 +15,19 @@ class NGraph : public GrowingNetwork {
 	public:
 
 		const long int DIM;
-		double radius;
-		double baseGam;
-		double baseTol;
-		double baseItr;
+		float radius;
+		float baseGam;
+		float baseTol;
+		float baseItr;
 		long int equalizationThreshold;	// the number of nodes below which GradientDescent is called once per node added
 		long int equalizationPeriod;	// the number of nodes added per call to GradientDescent
-		double iterationWeights;	// count of weighted iterations over this object's lifetime
+		float iterationWeights;	// count of weighted iterations over this object's lifetime
 		
 		NGraph(long int d);
 		SpatialVertex* getNode(const long int i) const;
-		virtual double* randomLocation() = 0;
-		virtual double linearDistance(Vertex* a, Vertex* b) = 0;
-		virtual double linearDistance(double* a, double* b) = 0;
+		virtual float* randomLocation() = 0;
+		virtual float linearDistance(Vertex* a, Vertex* b) = 0;
+		virtual float linearDistance(float* a, float* b) = 0;
 	
 };
